src: Remove clock mux buffers from FPGA implementation and dqsn delay
* dqs is not used as a differential signal when reading, so dqsn is
useless and not needed. We use it at write time to propagate the data to
the ram
* clock mux buffers are not needed when implementing the clock-based
multiplexer for double data rate (upper 16bit and lower 16 bits
alternatively), and are replaced by combinational logic most likely
implemented with LUTs. This is needed because we have constraints on the
db* location on the FMC card, and those clocks have to lie in the same
clock region, which can held only 12 clocks in total. It should work but
needs to be verified